pyproject.toml
[tool.poetry]
name = "chatbot-bedrock"
version = "0.1.0"
description = ""
authors = ["Trần Bá Hoàng <87892619+hoangbros03@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
streamlit = "^1.36.0"
boto3 = "^1.34.133"
langchain = "^0.2.6"
langchain-community = "^0.2.6"
python-dotenv = "^1.0.1"
langchain-aws = "^0.1.8"
opencv-python = "^4.10.0.84"
opencv-python-headless = "^4.10.0.84"
transformers = "^4.42.3"
anthropic = "^0.30.0"
pandas = "^2.2.2"
faiss-cpu = "^1.8.0.post1"
pypdf = "^4.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
